[Fix] DNS Flow Description
Change in the DNS Flow description to add the work of authoritative server as shown by arrow 6,7. 

Making this change since

A TLD nameserver is responsible for managing the top-level domain (TLD) portion of the DNS hierarchy. It holds information about the authoritative nameservers for domains that fall under its TLD.

An authoritative nameserver holds the actual DNS records (A, AAAA, MX, CNAME, etc.) for a specific domain. It provides the final answer for queries related to the domain name it manages.